pyproject.toml
[project]
name = "test_task"
version = "0.1.0"
description = ""
authors = [
    {name = "await.art@yandex.com"}
]
package-mode = false


[tool.poetry.dependencies]
python = ">=3.11,<3.13"
fastapi = "^0.115.0"
pydantic = "^2.8.2"
pydantic-settings = "^2.4.0"
uvicorn = {extras = ["standard"], version = "^0.30.6"}
SQLAlchemy = "^2.0.35"
psycopg = {version = "^3.2.1", extras=["binary"]}
python-dotenv = "^1.0.1"
asyncpg = "^0.30.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"
pytest-cov = "^5.0.0"
httpx = "^0.27.2"
ruff = "^0.6.8"
alembic = "^1.16.4"
asyncpg = "^0.30.0"
pytest-asyncio = "^1.1.0"
black = "^25.1.0"


[tool.ruff]
line-length = 79
select = ["E","F","I","UP","BLE","B","ARG","S","C90"]
ignore = ["S101", "B008"]

[tool.black]
line-length = 79


[tool.pytest.ini_options]
addopts = "-ra -q --cov=app --cov-report=term-missing"
pythonpath = ["."]


[build-system]
requires = ["poetry-core>=1.8.0"]
build-backend = "poetry.core.masonry.api"